+++ app/backend/src/main/kotlin/nito/club/backend/Application.kt
@@ -1,0 +1,3 @@
+package nito.club.backend
+
+fun main(args: Array<String>): Unit = io.ktor.server.netty.EngineMain.main(args)
