+++ app/backend/src/main/kotlin/nito/club/backend/graphql/query/HelloQuery.kt
@@ -1,0 +1,7 @@
+package nito.club.backend.graphql.query
+
+import com.expediagroup.graphql.server.operations.Query
+
+object HelloQuery : Query {
+    fun hello() = "Hello World!"
+}
