+++ app/backend/src/main/kotlin/nito/club/backend/graphql/server/KtorGraphQLContextFactory.kt
@@ -1,0 +1,6 @@
+package nito.club.backend.graphql.server
+
+import com.expediagroup.graphql.server.execution.GraphQLContextFactory
+import io.ktor.server.request.ApplicationRequest
+
+class KtorGraphQLContextFactory : GraphQLContextFactory<ApplicationRequest>
