+++ app/backend/src/main/kotlin/nito/club/backend/graphql/server/KtorGraphQLSchema.kt
@@ -1,0 +1,28 @@
+package nito.club.backend.graphql.server
+
+import com.expediagroup.graphql.generator.SchemaGeneratorConfig
+import com.expediagroup.graphql.generator.TopLevelObject
+import com.expediagroup.graphql.generator.scalars.IDValueUnboxer
+import com.expediagroup.graphql.generator.toSchema
+import graphql.GraphQL
+import nito.club.backend.graphql.query.HelloQuery
+
+object KtorGraphQLSchema {
+    private val config = SchemaGeneratorConfig(
+        supportedPackages = listOf("club.nito.graphql"),
+    )
+    private val queries = listOf(
+        TopLevelObject(HelloQuery),
+    )
+    private val mutations = listOf<TopLevelObject>()
+
+    private val graphQLSchema = toSchema(
+        config = config,
+        queries = queries,
+        mutations = mutations,
+    )
+
+    fun getGraphQLObject(): GraphQL = GraphQL.newGraphQL(graphQLSchema)
+        .valueUnboxer(IDValueUnboxer())
+        .build()
+}
