+++ app/backend/src/main/kotlin/nito/club/backend/schema/HelloQueryService.kt
@@ -1,0 +1,7 @@
+package nito.club.backend.schema
+
+import com.expediagroup.graphql.server.operations.Query
+
+class HelloQueryService : Query {
+    fun hello() = "World!"
+}
