+++ app/backend/src/main/resources/application.conf
@@ -1,0 +1,10 @@
+ktor {
+    deployment {
+        port = 8080
+    }
+    application {
+        modules = [
+            nito.club.backend.GraphQLModuleKt.graphQLModule
+        ]
+    }
+}
