--- bff/src/main/kotlin/nito/club/bff/Application.kt
@@ -1,3 +1,0 @@
-package nito.club.bff
-
-fun main(args: Array<String>): Unit = io.ktor.server.netty.EngineMain.main(args)
