--- bff/src/main/kotlin/nito/club/bff/graphql/query/HelloQuery.kt
@@ -1,7 +1,0 @@
-package nito.club.bff.graphql.query
-
-import com.expediagroup.graphql.server.operations.Query
-
-object HelloQuery : Query {
-    fun hello() = "Hello World!"
-}
