--- bff/src/main/kotlin/nito/club/bff/graphql/server/KtorGraphQLContextFactory.kt
@@ -1,6 +1,0 @@
-package nito.club.bff.graphql.server
-
-import com.expediagroup.graphql.server.execution.GraphQLContextFactory
-import io.ktor.server.request.ApplicationRequest
-
-class KtorGraphQLContextFactory : GraphQLContextFactory<ApplicationRequest>
