--- bff/src/main/kotlin/nito/club/bff/schema/BookQueryService.kt
@@ -1,18 +1,0 @@
-package nito.club.bff.schema
-
-import com.expediagroup.graphql.generator.annotations.GraphQLDescription
-import com.expediagroup.graphql.server.operations.Query
-import graphql.schema.DataFetchingEnvironment
-import nito.club.bff.schema.dataloaders.BookDataLoader
-import nito.club.bff.schema.models.Book
-import java.util.concurrent.CompletableFuture
-
-class BookQueryService : Query {
-    @GraphQLDescription("Return list of books based on BookSearchParameter options")
-    @Suppress("unused")
-    fun searchBooks(params: BookSearchParameters, dfe: DataFetchingEnvironment): CompletableFuture<List<Book>> =
-        dfe.getDataLoader<Int, Book>(BookDataLoader.dataLoaderName)
-            .loadMany(params.ids)
-}
-
-data class BookSearchParameters(val ids: List<Int>)
