--- bff/src/main/kotlin/nito/club/bff/schema/HelloQueryService.kt
@@ -1,7 +1,0 @@
-package nito.club.bff.schema
-
-import com.expediagroup.graphql.server.operations.Query
-
-class HelloQueryService : Query {
-    fun hello() = "World!"
-}
