--- bff/src/main/kotlin/nito/club/bff/schema/dataloaders/CourseDataLoader.kt
@@ -1,18 +1,0 @@
-package nito.club.bff.schema.dataloaders
-
-import com.expediagroup.graphql.dataloader.KotlinDataLoader
-import graphql.GraphQLContext
-import kotlinx.coroutines.runBlocking
-import nito.club.bff.schema.models.Course
-import org.dataloader.DataLoaderFactory
-import java.util.concurrent.CompletableFuture
-
-val CourseDataLoader = object : KotlinDataLoader<Int, Course?> {
-    override val dataLoaderName = "COURSE_LOADER"
-    override fun getDataLoader(graphQLContext: GraphQLContext) =
-        DataLoaderFactory.newDataLoader { ids ->
-            CompletableFuture.supplyAsync {
-                runBlocking { Course.search(ids).toMutableList() }
-            }
-        }
-}
