--- bff/src/main/kotlin/nito/club/bff/schema/dataloaders/UniversityDataLoader.kt
@@ -1,18 +1,0 @@
-package nito.club.bff.schema.dataloaders
-
-import com.expediagroup.graphql.dataloader.KotlinDataLoader
-import graphql.GraphQLContext
-import kotlinx.coroutines.runBlocking
-import nito.club.bff.schema.models.University
-import org.dataloader.DataLoaderFactory
-import java.util.concurrent.CompletableFuture
-
-val UniversityDataLoader = object : KotlinDataLoader<Int, University?> {
-    override val dataLoaderName = "UNIVERSITY_LOADER"
-    override fun getDataLoader(graphQLContext: GraphQLContext) =
-        DataLoaderFactory.newDataLoader { ids ->
-            CompletableFuture.supplyAsync {
-                runBlocking { University.search(ids).toMutableList() }
-            }
-        }
-}
