--- bff/src/main/resources/application.conf
@@ -1,10 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-    }
-    application {
-        modules = [
-            nito.club.bff.GraphQLModuleKt.graphQLModule
-        ]
-    }
-}
